pyproject.toml
[project]
name = "gpcore-sdk"
version = "0.1.0"
description = "Python SDK for the GPortal GPCore API, generated from protobuf specifications"
authors = [{ name = "Marc Seiler", email = "digitalfiz@proton.me" }]
readme = "README.md"
requires-python = ">=3.12"
repository = "https://github.com/digitalfiz/gpcore-sdk"
homepage = "https://github.com/digitalfiz/gpcore-sdk"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "grpcio-status>=1.60.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.12"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
# Ignore all checks on generated protobuf files
"src/gpcore_sdk/gpcore/**" = ["ALL"]
"src/gpcore_sdk/buf/**" = ["ALL"]
"src/gpcore_sdk/google/**" = ["ALL"]

